Fix update_todo_routes.R function bug
I noticed missing routes in the google-sheets to do list that were not
listed although I knew I had not climbed them. The problem arose when
the "attempted" column was computed (and subsequently filtered), as in
some cases a route was attempted by a single climber. One solution might
be to additionally distinguish attempts by the climber (append label).
I went for an "easier" solution by adding the missing routes back at the
end.